Create Enemy class
This class controls the spawning and movement of enemies. It also handles enemy behavior like being destroyed when hit by a bullet.
Enemy's movement: The enemies move downwards from the top of the screen. They are "alive" until hit by a bullet or a collision with the player.
Enemy's position: Enemies spawn randomly at the top of the screen and move downward.